Fixes for x64

+++ dbs/path.cpp
@@ -34,9 +34,9 @@ void dbs::Path::reverse()
 {
     if (nodes.size() < 2)
         return;
-    for (int i = 0, j = nodes.size() - 1; i < j; i++, j--)
+    for (size_t i = 0, j = nodes.size() - 1; i < j; i++, j--)
         std::swap(nodes[i].to_p(), nodes[j].to_p());
-    for (int i = 0, j = nodes.size() - 2; i < j; i++, j--)
+    for (size_t i = 0, j = nodes.size() - 2; i < j; i++, j--)
         std::swap(nodes[i].bulge, nodes[j].bulge);
     for (auto & node : nodes)
         node.bulge = -node.bulge;
